Refactor telemetry subsystem, naming things
